+++ src/main/java/com/bitTiger/searchAds/adsInfo/AdsInvertedIndex.java
@@ -6,24 +6,24 @@ import java.util.List;
 import java.util.Map;
 
 public class AdsInvertedIndex {
-  private Map<String, List<Integer>>  _adsInvertedIndex;
+  private Map<String, List<Long>>  _adsInvertedIndex;
 
   public AdsInvertedIndex() {
-    _adsInvertedIndex = new HashMap<String, List<Integer>>();
+    _adsInvertedIndex = new HashMap<String, List<Long>>();
   }
 
-  public void insertIndex(String keyWord, int adsId) {
+  public void insertIndex(String keyWord, long adsId) {
     if (!_adsInvertedIndex.containsKey(keyWord)) {
-      _adsInvertedIndex.put(keyWord, new ArrayList<Integer>());
+      _adsInvertedIndex.put(keyWord, new ArrayList<Long>());
     }
     _adsInvertedIndex.get(keyWord).add(adsId);
   }
 
-  public List<Integer> retrieveIndex(String keyWord) {
+  public List<Long> retrieveIndex(String keyWord) {
     return _adsInvertedIndex.get(keyWord);
   }
   
-  public Map<String, List<Integer>> GetAdsInvertedIndex()
+  public Map<String, List<Long>> GetAdsInvertedIndex()
   {
 	  return  _adsInvertedIndex;
   }
