+++ src/main/java/com/bitTiger/searchAds/adsInfo/CampaignInventory.java
@@ -4,21 +4,21 @@ import java.util.Map;
 import java.util.HashMap;
 
 public class CampaignInventory {
-  private final Map<Integer, CampaignInfo> _campaignInventory;
+  private final Map<Long, CampaignInfo> _campaignInventory;
 
   public CampaignInventory() {
-    _campaignInventory = new HashMap<Integer, CampaignInfo>();
+    _campaignInventory = new HashMap<Long, CampaignInfo>();
   }
 
   public void insertCampaign(CampaignInfo campaignInfo) {
     _campaignInventory.put(campaignInfo.getCampaignId(), campaignInfo);
   }
 
-  public CampaignInfo findCampaign(Integer campaignId) {
+  public CampaignInfo findCampaign(Long campaignId) {
     return _campaignInventory.get(campaignId);
   }  
   
-  public Map<Integer, CampaignInfo> GetCampaignInventory()
+  public Map<Long, CampaignInfo> GetCampaignInventory()
   {
 	  return _campaignInventory;
   }
